backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
manim==0.18.0
google-generativeai==0.3.2
python-multipart==0.0.9
python-dotenv==1.0.1